refactor-components * Move LockModal to components directory
