+++ src/components/LockModal/LockModal.tsx
@@ -3,13 +3,13 @@ import { Heading3, Paragraph } from '@entur/typography'
 import { Modal } from '@entur/modal'
 import { GridContainer, GridItem } from '@entur/grid'
 import { PrimaryButton } from '@entur/button'
-import Check from '../../../assets/images/check.png'
-import retinaCheck from '../../../assets/images/check@2x.png'
-import { useUser } from '../../../UserProvider'
-import { useSettings } from '../../../settings/SettingsProvider'
-import { CloseButton } from '../../CloseButton/CloseButton'
-import { LoginModal } from '../../LoginModal/LoginModal'
-import { LoginCase } from '../../LoginModal/login-modal-types'
+import Check from '../../assets/images/check.png'
+import retinaCheck from '../../assets/images/check@2x.png'
+import { useUser } from '../../UserProvider'
+import { useSettings } from '../../settings/SettingsProvider'
+import { CloseButton } from '../CloseButton/CloseButton'
+import { LoginModal } from '../LoginModal/LoginModal'
+import { LoginCase } from '../LoginModal/login-modal-types'
 import './LockModal.scss'
 
 interface LockModalProps {
